refactor(all module): refactor module factory

+++ multi-datasource/spring-boot-mybatis-multi-datasource/src/main/java/com/xiaodao/mybatismultidatasource/controller/UserController.java
@@ -1,8 +1,7 @@
 package com.xiaodao.mybatismultidatasource.controller;
 
-import com.xiaodao.mybatismultidatasource.model.User;
-import com.xiaodao.mybatismultidatasource.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.xiaodao.common.model.User;
+import com.xiaodao.service.UserService;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -18,19 +17,19 @@ import java.util.List;
 @RequestMapping("user")
 public class UserController {
 
-	private final UserService userService;
+    private final UserService userService;
 
-	public UserController(UserService userService) {
-		this.userService = userService;
-	}
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
 
-	@GetMapping
-	public List<User> getUserById(@RequestParam(required = false, defaultValue = "1") Byte age) {
-		return userService.findUserByAgeWithin(age);
-	}
+    @GetMapping
+    public List<User> getUserById(@RequestParam(required = false, defaultValue = "1") Integer age) {
+        return userService.findUserByAgeWithin(age);
+    }
 
-	@PostMapping
-	public User createUser(@RequestBody User user) {
-		return userService.saveUser(user);
-	}
+    @PostMapping
+    public User createUser(@RequestBody User user) {
+        return userService.saveUser(user);
+    }
 }
